tests/wpt/web-platform-tests/largest-contentful-paint/image-inside-svg.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: observe image inside SVG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
setup({"hide_test_state": true});
async_test(function (t) {
  assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
  const beforeLoad = performance.now();
  new PerformanceObserver(
    t.step_func_done(entryList => {
      assert_equals(entryList.getEntries().length, 1);
      const entry = entryList.getEntries()[0];
      const url = window.location.origin + '/images/blue.png';
      // blue.png is 133 by 106.
      const size = 133 * 106;
      checkImage(entry, url, 'image_id', size, beforeLoad);
    })
  ).observe({type: 'largest-contentful-paint', buffered: true});
}, "Image inside SVG is observable.");
</script>
<svg width="300" height="300" id='svg_id'>
  <image href='/images/blue.png' id='image_id'/>
</svg>
